Create Project 1
The task is to calculate the total number of working hours between two given dates, excluding all Sundays and the 1st and 2nd Saturdays. However, if either of the input dates falls on a Sunday or the 1st or 2nd Saturday, that particular date should still be included in the calculation.